requirements.txt
# 牛肉マーブリング判定システム PWA版
# 依存関係一覧

# ===== Webフレームワーク =====
Flask==2.3.3
Flask-CORS==4.0.0

# ===== 機械学習・AI =====
tensorflow==2.13.0
numpy==1.24.3
Pillow==10.0.0
opencv-python==4.8.0.76

# ===== データ処理 =====
pandas==2.0.3

# ===== Webサーバー（本番環境用） =====
gunicorn==21.2.0
waitress==2.1.2

# ===== 開発・デバッグ用 =====
python-dotenv==1.0.0

# ===== セキュリティ =====
Werkzeug==2.3.7

# ===== ユーティリティ =====
pathlib2==2.3.7;python_version<"3.4"

# ===== 型ヒント（開発用） =====
typing-extensions==4.7.1

# ===== JSON処理 =====
jsonschema==4.19.0

# ===== 日時処理 =====
python-dateutil==2.8.2

# ===== ログ処理 =====
colorlog==6.7.0

# ===== ファイル処理 =====
send2trash==1.8.2

# ===== システム情報 =====
psutil==5.9.5 